Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

Responding to your comments in the Commit message.  In the future, just ask questions in the patch set comments (here), instead of the commit message.  It makes the commit history a bit cleaner :)

1. Definitely true, good catch, this has the potential to cause logins to fail.
2. A good choice.
3. Good idea, server is in Python, where the hexdigest method always provides lowercase strings, as does the CMAccountWebsite project.
4. I don't foresee this causing any issues.  When you call getBytes() on a String, it uses the default charset, which is UTF-8 on Android.  See http://developer.android.com/reference/java/lang/String.html#getBytes().

I'm going to go ahead and merge this.

Patch-set: 1
Label: Verified=+1
Label: Code-Review=+2