Update patch set 1
Patch Set 1:

Thanks Chris,

Appreciate the info regarding comments, 1st commit here - still learning best practices :)

regarding #4 - i wasn't sure if there is a default one in cm, it's good to know that it's utf-8. 
just to elaborate on the point - i was thinking about a situation where someone might have registered on the site with some specific encoding, and then uses the phone with another encoding (or the default one) which might be different.
i have not had the chance to look at the website code yet (i think i read that it was also released) but it might be good to specify something generic both there and in the code (like utf-8), that would handle most chars or restrict the type of chars during registration (if that's not already done).

was trying to fit this in a "short as possible commit message"...

Patch-set: 1
Reviewer: Tom Keidar <6879@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0