Update patch set 1
Patch Set 1:

i was not referring to the email in any way - all of my comments relate to the password (hashed part).

i was referring to a situation where the encoding on the phone is different then the one in the browser. then, a password encoded in hebrew (e.g - i know people who do that to make stronger passwords) and then hashed, will not be compatible with that same password (typed with the heb keyboard on android) but encoded in utf-8 and then hashed.
the situation for a browser being default set for a different encoding is pretty widespread in non-english speaking countries, and i have encountered this on many other projects.

i apologize if i could not make that clear before (was not referring to the email in any way), my suggestion was to either always force utf-8 by explicitly specifying (or something else you deem cross-compatible), or force password restrictions (which a lot of sites do).

Patch-set: 1