Update patch set 1
Patch Set 1:

Ah okay, that makes more sense.  The JavaScript side of the website is open source, the project is CMAccountWebsite.  Feel free to submit a change that addresses this problem, and I'll give it a review.

I'm not too concerned about fixing this at the moment, as I've got a full plate of security related enhancements to make.  If this does become an issue, it would definitely be a very small edge case.

Patch-set: 1