Update patch set 1
Patch Set 1:

yeah, i do not think this will become an issue, which is why i did not include it in the original commit, but as a point for your consideration (though, i have been surprised in the past by the amounts of people using russian/german/chinese/korean/japanese encodings ;) ).

I will try and take a look at the javascript code as you suggested. REALLY appreciate what you guys are bringing, and will be happy to help with any security enhancements/features (have a lot of experience in this), is there an open todo / bug list for this somewhere, that i can  contribute? or are you guys not at the point you need more hands yet?

Patch-set: 1