Fix for race condition after Google account setup.
The future for add account can be delayed enough that it is possible to move on to the last page by the time it is called back. The fix is two-fold. Added a check to ensure we don't call getPage on an out of bounds index. Also changed to call into account manager to return the Intent for Google setup. Now we can use onActivityResult and avoid this race condition.

Change-Id: I3bceb7b154a74e34bcc1a27331cccc57e0d9b573