Update patch set 1
Patch Set 1:

> You should add lineage copyright only to the files you modified,
 > not all of them
 > Also don't update cyanogenmod copyright

Why should I only change it on files I modified? Alo, if I don't change CyanogenMod copyright, this wouldn't be Copyrighted from ~2013 to 2016. What's so bad if I append the years of Copyright until CM died? I only know that it is bad if I Remove the CyanogenMod Copyright header and replace it.

Patch-set: 1